feat: Add last 10 previous orders
